refactor: map/filter 페이지 component 정리
