+++ client/map/filter/ExtraFilter.tsx
@@ -1,61 +1,66 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
+
 import CategoryHeader2 from 'components/common/CategoryHeader2';
-import useMapFilterRedux from 'hooks/useMapFilterRedux';
 import {
-  dispatchFloorfive,
-  dispatchFloorsixup,
-  dispatchFloortenup,
-  dispatchBasement,
-  dispatchRooftop,
+  dispatchCar,
+  dispatchElevator,
+  dispatchPet,
+  dispatchFullOption,
+  dispatchWomen,
+  dispatchLoan,
 } from 'store/modules/filter';
-import { RootState } from '../../store/modules/index';
 import CommonBtn from './CommonBtn';
 
-export default function Floor() {
-  const [state, dispatch] = useMapFilterRedux();
-
+function TransactionType() {
   const [btn1Check, setBtn1Check] = useState(false);
   const [btn2Check, setBtn2Check] = useState(false);
   const [btn3Check, setBtn3Check] = useState(false);
   const [btn4Check, setBtn4Check] = useState(false);
   const [btn5Check, setBtn5Check] = useState(false);
+  const [btn6Check, setBtn6Check] = useState(false);
 
   const btnArr = [
     {
       check: btn1Check,
       checkHandler: setBtn1Check,
-      type: '5층 이하',
-      dispatch: dispatchFloorfive,
+      type: '주차가능',
+      dispatch: dispatchCar,
     },
     {
       check: btn2Check,
       checkHandler: setBtn2Check,
-      type: '6층~10층',
-      dispatch: dispatchFloorsixup,
+      type: '엘리베이터',
+      dispatch: dispatchElevator,
     },
     {
       check: btn3Check,
       checkHandler: setBtn3Check,
-      type: '10층 이상',
-      dispatch: dispatchFloortenup,
+      type: '반려동물',
+      dispatch: dispatchPet,
     },
     {
       check: btn4Check,
       checkHandler: setBtn4Check,
-      type: '반지하',
-      dispatch: dispatchBasement,
+      type: '풀옵션',
+      dispatch: dispatchFullOption,
     },
     {
       check: btn5Check,
       checkHandler: setBtn5Check,
-      type: '옥탑방',
-      dispatch: dispatchRooftop,
+      type: '여성전용',
+      dispatch: dispatchWomen,
+    },
+    {
+      check: btn6Check,
+      checkHandler: setBtn6Check,
+      type: '전세대출',
+      dispatch: dispatchLoan,
     },
   ];
 
   return (
-    <div className="border-t border-border_color">
-      <CategoryHeader2 title="층 수" />
+    <div className="border-border_color border-t">
+      <CategoryHeader2 title="추가필터" />
       <div className="w-full grid grid-cols-3 gap-room_register_gap">
         {btnArr.map((item, index) => (
           <CommonBtn
@@ -68,6 +73,11 @@ export default function Floor() {
           />
         ))}
       </div>
+      <div className="text-[12px] Pretendard-Regular text-font_gray my-6">
+        * 풀옵션 항목은 세탁기, 냉장고, 에어컨, 가스레인지입니다.
+      </div>
     </div>
   );
 }
+
+export default TransactionType;
