+++ client/map/filter/RoomType.tsx
@@ -1,17 +1,13 @@
-import { useEffect, useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import CategoryHeader2 from 'components/common/CategoryHeader2';
-import useMapFilterRedux from 'hooks/useMapFilterRedux';
+import { useState } from 'react';
+
 import {
-  dispatchOpenType,
-  dispatchSeparateType,
-  dispatchTwoFloorType,
+  dispatchOneRoom,
+  dispatchTwoRoom,
+  dispatchThreeRoom,
 } from 'store/modules/filter';
 import CommonBtn from './CommonBtn';
 
-export default function TransactionType() {
-  const [state, dispatch] = useMapFilterRedux();
-
+function TransactionType() {
   const [btn1Check, setBtn1Check] = useState(false);
   const [btn2Check, setBtn2Check] = useState(false);
   const [btn3Check, setBtn3Check] = useState(false);
@@ -20,26 +16,31 @@ export default function TransactionType() {
     {
       check: btn1Check,
       checkHandler: setBtn1Check,
-      type: '오픈형',
-      dispatch: dispatchOpenType,
+      type: '원룸',
+      dispatch: dispatchOneRoom,
     },
     {
       check: btn2Check,
       checkHandler: setBtn2Check,
-      type: '분리형',
-      dispatch: dispatchSeparateType,
+      type: '투룸',
+      dispatch: dispatchTwoRoom,
     },
     {
       check: btn3Check,
       checkHandler: setBtn3Check,
-      type: '복층형',
-      dispatch: dispatchTwoFloorType,
+      type: '쓰리룸',
+      dispatch: dispatchThreeRoom,
     },
   ];
 
   return (
-    <>
-      <CategoryHeader2 title="방 구조" />
+    <div className="mt-[30px] pb-[30px] border-b border-border_color">
+      <div className="pb-6 text-[17px] Pretendard-SemiBold flex items-center">
+        방 종류
+        <span className="ml-2 text-[12px] Pretendard-Regular text-font_gray">
+          &nbsp;중복 선택 가능
+        </span>
+      </div>
       <div className="w-full grid grid-cols-3 gap-room_register_gap">
         {btnArr.map((item, index) => (
           <CommonBtn
@@ -52,6 +53,8 @@ export default function TransactionType() {
           />
         ))}
       </div>
-    </>
+    </div>
   );
 }
+
+export default TransactionType;
