+++ client/map/filter/Structure.tsx
@@ -1,48 +1,42 @@
 import { useState } from 'react';
-import useMapFilterRedux from 'hooks/useMapFilterRedux';
+
+import CategoryHeader2 from 'components/common/CategoryHeader2';
 import {
-  dispatchOneRoom,
-  dispatchTwoRoom,
-  dispatchThreeRoom,
+  dispatchOpenType,
+  dispatchSeparateType,
+  dispatchTwoFloorType,
 } from 'store/modules/filter';
 import CommonBtn from './CommonBtn';
 
-export default function TransactionType() {
+function TransactionType() {
   const [btn1Check, setBtn1Check] = useState(false);
   const [btn2Check, setBtn2Check] = useState(false);
   const [btn3Check, setBtn3Check] = useState(false);
-
-  const [state, dispatch] = useMapFilterRedux();
 
   const btnArr = [
     {
       check: btn1Check,
       checkHandler: setBtn1Check,
-      type: '원룸',
-      dispatch: dispatchOneRoom,
+      type: '오픈형',
+      dispatch: dispatchOpenType,
     },
     {
       check: btn2Check,
       checkHandler: setBtn2Check,
-      type: '투룸',
-      dispatch: dispatchTwoRoom,
+      type: '분리형',
+      dispatch: dispatchSeparateType,
     },
     {
       check: btn3Check,
       checkHandler: setBtn3Check,
-      type: '쓰리룸',
-      dispatch: dispatchThreeRoom,
+      type: '복층형',
+      dispatch: dispatchTwoFloorType,
     },
   ];
 
   return (
-    <div className="mt-[30px] pb-[30px] border-b border-border_color">
-      <div className="pb-6 text-[17px] Pretendard-SemiBold flex items-center">
-        방 종류
-        <span className="ml-2 text-[12px] Pretendard-Regular text-font_gray">
-          &nbsp;중복 선택 가능
-        </span>
-      </div>
+    <>
+      <CategoryHeader2 title="방 구조" />
       <div className="w-full grid grid-cols-3 gap-room_register_gap">
         {btnArr.map((item, index) => (
           <CommonBtn
@@ -55,6 +49,8 @@ export default function TransactionType() {
           />
         ))}
       </div>
-    </div>
+    </>
   );
 }
+
+export default TransactionType;
